+++ packages/core/src/format/path.ts
@@ -47,36 +47,98 @@ export function usePath<T = any, TSource = any>(
     let c: Record<string, any> = s;
     for (let i = 0; i < fragments.length; i++) {
       let fragmentPath = fragments[i];
-      let index: any = -1;
 
       if (fragmentPath[fragmentPath.length - 1] === "]") {
-        const m = fragmentPath.match(/\[(\d+)\]$/);
-        if (m && m[1]) {
-          index = +m[1];
+        const r = /\[[`'"]?([^`'"\]]*)[`'"]?\]/g;
+        let path = fragmentPath;
+        let m = r.exec(path);
 
-          fragmentPath = fragmentPath.slice(0, -m[0].length);
+        if (m) {
+          let lastLen = m[0].length;
+          let lastIndex = m.index - lastLen;
+          let mi = 1;
+
+          do {
+            if (lastIndex + lastLen !== m.index) {
+              // istanbul ignore else
+              if (__DEV__) {
+                console.warn(`[usePath] invalid path "${fragments[i]}"`);
+              }
+            }
+            lastIndex = m.index;
+            lastLen = m[0].length;
+
+            fragmentPath = fragmentPath.slice(0, -m[0].length);
+            fragments.splice(i + mi, 0, m[1]);
+
+            ++mi;
+          } while ((m = r.exec(path)));
+
+          // if the fragmentPath is empty, eg: [1][1]
+          // we should continue until the next path
+          if (!fragmentPath && path[0] === "[" && path.length > 2) {
+            continue;
+          }
+        } else {
+          fragmentPath = "";
+          console.warn(`[usePath] invalid path provided "${path}"`);
         }
       }
 
       if (isObject(c)) {
-        c = c[fragmentPath];
-
-        // array like: when using ref with and array, it becomes arraylike object
-        if (index >= 0) {
-          c = (c as any)[index];
+        if (!fragmentPath) {
+          // istanbul ignore else
+          if (__DEV__) {
+            console.warn(
+              `Path "${fragments
+                .slice(0, i + 1)
+                .join(separator)}" doesn't exist on:`,
+              source
+            );
+          }
+          return notFoundReturn(
+            fragments.slice(0, i + 1).join(separator),
+            c,
+            p,
+            s
+          );
         }
+
+        c = c[fragmentPath];
       } else {
+        // istanbul ignore else
         if (__DEV__) {
           console.warn(
-            `Path "${fragments.slice(0, i).join(separator)}" doesn't exist on:`,
+            `Path "${fragments
+              .slice(0, i + 1)
+              .join(separator)}" doesn't exist on:`,
             source
           );
         }
-        return notFoundReturn(fragments.slice(0, i).join(separator), c, p, s);
+        return notFoundReturn(
+          fragments.slice(0, i + 1).join(separator),
+          c,
+          p,
+          s
+        );
       }
 
       if (!c) {
-        return notFoundReturn(fragments.slice(0, i).join(separator), c, p, s);
+        // istanbul ignore else
+        if (__DEV__) {
+          console.warn(
+            `Path "${fragments
+              .slice(0, i + 1)
+              .join(separator)}" doesn't exist on:`,
+            source
+          );
+        }
+        return notFoundReturn(
+          fragments.slice(0, i + 1).join(separator),
+          c,
+          p,
+          s
+        );
       }
     }