+++ packages/core/src/promise/index.ts
@@ -1,3 +1,4 @@
 export * from "./promise";
 export * from "./cancellablePromise";
 export * from "./retry";
+export * from "./promiseLazy";