+++ packages/core/src/promise/promise.ts
@@ -1,5 +1,5 @@
 import { ref, Ref } from "@vue/composition-api";
-import { isBoolean } from "../utils";
+import { isBoolean, isObject } from "../utils";
 
 type PromiseType<T extends Promise<any>> = T extends Promise<infer R>
   ? R
@@ -24,14 +24,31 @@ export interface PromiseResultFactory<
   exec: (...args: TArgs) => Promise<PromiseType<T> | undefined>;
 }
 
+export interface PromiseOptions {
+  /**
+   * if `true` allows to catch exception when `exec()`
+   */
+  throwException?: boolean;
+
+  /**
+   * Only executes on `exec`
+   */
+  lazy?: boolean;
+}
+
 /**
  *
  * @param fn - factory function
- * @param throwException - if `true` allows to catch exception when `exec()`
+ * @param lazy - Delays execution until `exec` is called
  */
 export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
   fn: (...args: TArgs) => Promise<T>,
-  throwException?: boolean
+  lazy?: boolean
+): PromiseResultFactory<Promise<T>, TArgs>;
+
+export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
+  fn: (...args: TArgs) => Promise<T>,
+  options: PromiseOptions
 ): PromiseResultFactory<Promise<T>, TArgs>;
 
 export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
@@ -40,7 +57,12 @@ export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
 
 export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
   fn: (...args: TArgs) => T,
-  throwException: boolean
+  lazy: boolean
+): PromiseResultFactory<Promise<T>, TArgs>;
+
+export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
+  fn: (...args: TArgs) => T,
+  options: PromiseOptions
 ): PromiseResultFactory<Promise<T>, TArgs>;
 
 export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
@@ -49,7 +71,12 @@ export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
 
 export function usePromise<T = any>(
   fn: () => Promise<T>,
-  throwException: boolean
+  lazy: boolean
+): PromiseResultFactory<Promise<T>>;
+
+export function usePromise<T = any>(
+  fn: () => Promise<T>,
+  options: PromiseOptions
 ): PromiseResultFactory<Promise<T>>;
 
 export function usePromise<T = any>(
@@ -58,7 +85,12 @@ export function usePromise<T = any>(
 
 export function usePromise<T = any>(
   fn: () => T,
-  throwException: boolean
+  lazy: boolean
+): PromiseResultFactory<Promise<T>>;
+
+export function usePromise<T = any>(
+  fn: () => T,
+  options: PromiseOptions
 ): PromiseResultFactory<Promise<T>>;
 
 export function usePromise<T = any>(
@@ -67,7 +99,7 @@ export function usePromise<T = any>(
 
 export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
   fn: (...args: TArgs) => T,
-  throwException = false
+  lazyOptions?: PromiseOptions | boolean
 ): PromiseResultFactory<T, TArgs> {
   if (!fn) {
     throw new Error(`[usePromise] argument can't be '${fn}'`);
@@ -75,6 +107,12 @@ export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
   if (typeof fn !== "function") {
     throw new Error(`[usePromise] expects function, but received ${typeof fn}`);
   }
+
+  const [lazy, throwException] = isBoolean(lazyOptions)
+    ? [lazyOptions, false]
+    : isObject(lazyOptions)
+    ? [lazyOptions.lazy, lazyOptions.throwException]
+    : [false, false];
 
   const loading = ref(false);
   const error = ref(null);
@@ -84,9 +122,8 @@ export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
   const exec = async (...args: TArgs): Promise<PromiseType<T> | undefined> => {
     loading.value = true;
     error.value = null;
-    result.value = null;
 
-    let throwExp =
+    const throwExp =
       args &&
       fn.length !== args.length &&
       args.length > 0 &&
@@ -113,6 +150,22 @@ export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
       }
     }
   };
+
+  if (!lazy) {
+    /* istanbul ignore else */
+    if (__DEV__) {
+      if (
+        fn.length > 0 &&
+        !isBoolean(isObject(lazyOptions) ? lazyOptions.lazy : lazyOptions)
+      ) {
+        console.warn(
+          "[usePromise] parameters detected on `fn` factory. Executing promise without arguments."
+        );
+      }
+    }
+    // @ts-ignore
+    exec();
+  }
 
   return {
     exec,